Prevent promotions when price changes less than 5%
The package is now passing an eighth test that prevents price changes
of under 5% from impacting promotion status. In addition, the code now
correctly rounds currency at the time of construction and setting. Time
permitting, it would be best to add a new price class to keep Product
focused on its primary task.