Prevent new promo start within 30 days of previous
We are now part way to meeting the final condition for the red pencil
kata. The package will no longer start a new promo within 30 days of
the last one. Now we just need to make sure that it *will* start a new
promo after that 30 day window.

A side effect of the implementation is that the code now has a logical
way of handling promotion eligibility within the first 30 days after
product creation. Namely, it now treats the creation time as it would
any other time when a price is set and, hence, prevents promos from
occurring for the first 30 days. This has the advantage of preventing
merchants from exploiting Red Pencil Promotions in yet another way (by
repeatedly creating duplicate products and immediately lowering prices
to keep their goods perpetually promoted).