Replace promotionStartDate usages with Promotions
I've now started to incorporate the Promotion class into the
implementation of Product. As a first step we have added a Promotion
variable (promotion) in the Product class and used it to get rid of the
promotionStartDate variable. This seemed like a natural starting point
and gave us an excuse to create some initial getters and setters for the
Promotion class.

I continue to practice writing as little code as possible to achieve
green tests during this final refactoring stage. Accordingly some of the
usages of Promotion getters will likely be expanded on in future
commits.