Simplify and streamline setPrice and isPromoted
Using the isActive method from the Promotion class, we may
simultaneously reduce the length of the setPrice and isPromoted
functions while also making their behavior more logical and improving
readability. As a part of this change, the logic required to start and
end promotions when prices change is now encapsulated in a private
updatePromotionStatus helper function, and Product methods no longer
rely on an isPromoted flag.